+++ src/app/Library/CrudPanel/CrudField.php
@@ -237,6 +237,25 @@ class CrudField
     }
 
     /**
+     * When subfields are defined, pass them through the guessing function
+     * so that they have label, relationship attributes, etc.
+     *
+     * @param  array  $subfields  Subfield definition array
+     * @return self
+     */
+    public function subfields($subfields)
+    {
+        $callAttributeMacro = ! isset($this->attributes['subfields']);
+        $this->attributes['subfields'] = $subfields;
+        $this->attributes = $this->crud()->makeSureFieldHasNecessaryAttributes($this->attributes);
+        if ($callAttributeMacro) {
+            $this->callRegisteredAttributeMacros();
+        }
+
+        return $this->save();
+    }
+
+    /**
      * Save the validation rules on the CrudPanel per field basis.
      *
      * @param  string  $rules  the field rules: required|min:1|max:5
